+++ Journal.Core/Exceptions/IncompatibleBackupException.cs
@@ -1,0 +1,9 @@
+namespace Journal.Exceptions
+{
+    public class IncompatibleBackupException : Exception
+    {
+        public IncompatibleBackupException(string message) : base(message)
+        {
+        }
+    }
+}